+++ packages/@react-aria/utils/src/index.ts
@@ -34,3 +34,4 @@ export {useValueEffect} from './useValueEffect';
 export {scrollIntoView, scrollIntoViewport} from './scrollIntoView';
 export {clamp, snapValueToStep} from '@react-stately/utils';
 export {isVirtualClick, isVirtualPointerEvent} from './isVirtualEvent';
+export {useEffectEvent} from './useEffectEvent';
