+++ packages/@react-stately/utils/src/index.ts
@@ -11,4 +11,3 @@
  */
 export {useControlledState} from './useControlledState';
 export {clamp, snapValueToStep, toFixedNumber} from './number';
-export {useEffectEvent} from './useEvent';
